added base for cli

+++ examples/ftgo/pkg/order/domain/order/order_line_item.go
@@ -5,7 +5,7 @@ import (
 )
 
 type OrderLineItem struct {
-	Quantity   int64  `json:"quantity"`
+	Quantity   int    `json:"quantity"`
 	MenuItemId string `json:"menu_item_id"`
 }
